refactor: move blinkstick class to new client package
