+++ tests/clients/test_blinkstick_matrix.py
@@ -1,4 +1,4 @@
-from blinkstick.blinkstick import BlinkStickProMatrix
+from blinkstick.clients.blinkstick import BlinkStickProMatrix
 
 
 def test_instantiate():
